specs: split out GetClusterNamespace()
